moved hasPermissions to the config.ts file, pluralise-ish actionResultNoAccess's message parameter
